Completed posts view, navigation between older posts and newer posts and url redirection to posts and pages
url - url no longer has to have a page number. pagination service sees to it that if there are less posts than required per page it changes the url to no number.
typing the url of a post will automatically redirect to the disred post.

postsService - added a local data base to store all the loaded posts for reuse

older and newer work as expected

small refactoring