SideBar clickable filters are working.
Added new function runQuery that runs every rout reload. it checks the current location and if there are any query parameters and runs the parameters found (if they match a predefined set of values) against a switch case. switch case then runs filter functions depending on the category
found or defaults to an 0 if no case found.
the cases are :
 - category
 - author
 - month-year
 - search ( not yet available )